backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
Pillow>=10.2.0
imagehash==4.3.1
piexif==1.1.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
watchdog==3.0.0

# EXIF and Metadata
exifread==3.0.0

# Additional
pydantic==2.5.3